# Small set of follow-up optimizations for _JTDAJ