* value-gating the atomic add for 0 values
* move nefc lookup out of the loop
* make dim_X and nblocks_per_block static to enable unrolling the loop

Difficult to tell the effect of individual parts here, but mostly better codegen and overall some perf gains (humanoid.xml):

before:
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.66 s
 Total simulation time: 4.46 s
 Total steps per second: 1,834,768
 Total realtime factor: 9,173.84 x
 Total time per step: 545.03 ns
```

after: 
```
Summary for 8192 parallel rollouts

 Total JIT time: 0.66 s
 Total simulation time: 4.35 s
 Total steps per second: 1,884,647
 Total realtime factor: 9,423.23 x
 Total time per step: 530.60 ns
```